Pull out codegen utilities to break a cyclic dependency
